Web Application/requirements.txt
black==23.3.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Flask==2.3.2
fonttools==4.39.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
threadpoolctl==3.1.0
tzdata==2023.3
Werkzeug==2.3.5
